inference/requirements.txt
flask
joblib
openai-whisper
torch
scikit-learn
prometheus_client